Common: Good looking dpad

+++ include/combo.h
@@ -134,7 +134,7 @@ void comboSetObjectSegment(GfxContext* gfx, void* buffer);
 void comboDrawObject(GameState_Play* play, Actor* actor, u16 objectId, u16 shaderId, int flags);
 void comboDrawGI(GameState_Play* play, Actor* actor, int gi, int flags);
 void comboDrawInit2D(GameState_Play* play);
-void comboDrawBlit2D(GameState_Play* play, u32 segAddr, int w, int h, int x, int y, float scale);
+void comboDrawBlit2D(GameState_Play* play, u32 segAddr, int w, int h, float x, float y, float scale);
 
 /* Event */
 void comboOotSetEventChk(u16 flag);
